* winsup.api/crlf.c: Fix C signed/unsigned compare warning. 	* winsup.api/mmaptest01.c: Ditto. 	* winsup.api/ltp/chmod01.c: Ditto. 	* winsup.api/ltp/fork04.c: Ditto. 	* winsup.api/ltp/lseek03.c: Ditto. 	* winsup.api/ltp/lseek06.c: Ditto. 	* winsup.api/ltp/lseek07.c: Ditto. 	* winsup.api/ltp/lseek08.c: Ditto. 	* winsup.api/ltp/mmap001.c: Ditto. 	* winsup.api/ltp/mmap02.c: Ditto. 	* winsup.api/ltp/mmap03.c: Ditto. 	* winsup.api/ltp/mmap04.c: Ditto. 	* winsup.api/ltp/mmap05.c: Ditto. 	* winsup.api/ltp/mmap06.c: Ditto. 	* winsup.api/ltp/mmap07.c: Ditto. 	* winsup.api/ltp/mmap08.c: Ditto. 	* winsup.api/ltp/pipe11.c: Ditto. 	* winsup.api/ltp/poll01.c: Ditto. 	* winsup.api/ltp/sync02.c: Ditto. 	* winsup.api/ltp/times03.c: Ditto. 	* winsup.api/ltp/umask03.c: Ditto. 	* winsup.api/ltp/getpgid01.c: Remove unused obsolete include. 	* winsup.api/ltp/getpgid02.c: Ditto.

+++ winsup/testsuite/winsup.api/crlf.c
@@ -156,7 +156,7 @@ int commands[] = {
 
 int errors = 0;
 
-#define num_commands (sizeof(commands)/sizeof(commands[0]))
+#define num_commands (int)(sizeof(commands)/sizeof(commands[0]))
 
 int pc;
 int askfor, get, expect, count, posn, whence, size;
